replace sidebar with drawer controller
replace sidebar with drawer controller

+++ SwipeProject/SwipeProject/Libs/AudioPlayer/AudioStreamer.m
@@ -584,7 +584,7 @@ void ASReadStreamCallBack
 		if (fileLength > 0 && seekByteOffset > 0)
 		{
 			CFHTTPMessageSetHeaderFieldValue(message, CFSTR("Range"),
-				(CFStringRef)[NSString stringWithFormat:@"bytes=%ld-%ld", seekByteOffset, fileLength]);
+				(CFStringRef)[NSString stringWithFormat:@"bytes=%ld-%ld", (long)seekByteOffset, (long)fileLength]);
 			discontinuous = YES;
 		}
